makzen-collaborative-filter/requirements.txt
flask

pandas


sortedcontainers

mysql.connector

flask_cors